+++ source/textscreen/txt_radiobutton.c
@@ -34,7 +34,7 @@ static void TXT_RadioButtonSizeCalc(TXT_UNCAST_ARG(radiobutton))
 
     // Minimum width is the string length + two spaces for padding
 
-    radiobutton->widget.w = strlen(radiobutton->label) + 6;
+    radiobutton->widget.w = (unsigned)(strlen(radiobutton->label) + 6);
     radiobutton->widget.h = 1;
 }
 
@@ -74,7 +74,7 @@ static void TXT_RadioButtonDrawer(TXT_UNCAST_ARG(radiobutton), int selected)
 
     TXT_DrawString(radiobutton->label);
     
-    for (i=strlen(radiobutton->label); i < w-6; ++i)
+    for (i=(int)strlen(radiobutton->label); i < w-6; ++i)
     {
         TXT_DrawString(" ");
     }
